Fix remaining SEO audit follow-ups
