+++ src/lib/seo/sitemapData.ts
@@ -5,7 +5,7 @@
  * We split into two child sitemaps so Google can segment crawl/lastmod
  * tracking per content type:
  *   - `/sitemap-pages.xml`  — core site (home, academic, courses, STEAM, camps)
- *   - `/sitemap-blogs.xml`  — blog posts + resource guides
+ *   - `/sitemap-blogs.xml`  — blog posts + resource guide articles
  * Index at `/sitemap.xml` references both (generated by `src/app/sitemap.ts`).
  */
 
@@ -253,12 +253,6 @@ export function buildPagesPaths(): string[] {
 export function buildBlogUrls(baseUrl: string, lastmod: string): SitemapUrl[] {
   const urls: SitemapUrl[] = []
   locales.forEach((locale) => {
-    urls.push({
-      loc: absoluteSiteUrl(RESOURCES_PATH, locale, baseUrl),
-      lastmod,
-      changefreq: 'weekly',
-      priority: 0.85,
-    })
     blogPostPaths.forEach((path) => {
       urls.push({
         loc: absoluteSiteUrl(path, locale, baseUrl),
@@ -284,7 +278,6 @@ export function buildBlogUrls(baseUrl: string, lastmod: string): SitemapUrl[] {
 export function buildBlogPaths(): string[] {
   const paths: string[] = []
   locales.forEach((locale) => {
-    paths.push(publicPath(RESOURCES_PATH, locale))
     blogPostPaths.forEach((path) => {
       paths.push(publicPath(path, locale))
     })
